requirements.txt
crewai==0.11.0
python-dotenv==1.0.0
pyyaml==6.0.1
langchain==0.1.0